Commence webiste
I got completely zoned out while making this website that I simply
forgot Git commits were a thing at all. Hence the lack of good commit
messages on this project so far. On the other hand, since HTML + CSS
alone are not that hard to follow along, it is ok I guess. Introduced
containers for the header, menu and quote.